+++ lib/base_interactor.rb
@@ -62,6 +62,73 @@ class BaseInteractor
     changeset.to_h.merge(updated_by: @user.user_name)
   end
 
+  # Remove all parameters for an +extended_columns+ field from a normal params object.
+  # Return the params and the extended_columns separately.
+  #
+  # @param params [Hash] the request parameters.
+  # @return [Array] the params and extended_columns Hashes.
+  def unwrap_extended_columns_params(params)
+    parms = {}
+    ext = {}
+    params.each do |name, value|
+      if name.to_s.start_with?('extcol_')
+        ext[name.to_s.delete_prefix('extcol_').to_sym] = value
+      else
+        parms[name] = value
+      end
+    end
+    [parms, ext]
+  end
+
+  # Return all parameters for an +extended_columns+ field from a params object.
+  # The keys of the resulting hash can optionally retain the "extcol_" prefix or have it stripped off.
+  #
+  # @param params [Hash] the request parameters.
+  # @param delete_prefix [boolean] should the "extcol_" prefix be removed from the keys? Default true.
+  # @return [Hash] the extended_columns Hash.
+  def select_extended_columns_params(params, delete_prefix: true)
+    selection = params.select { |a| a.to_s.start_with?('extcol_') }
+    if delete_prefix
+      selection.transform_keys { |k| k.to_s.delete_prefix('extcol_').to_sym }
+    else
+      selection
+    end
+  end
+
+  # Apply validation rules to a set of extended_columns and return the results.
+  #
+  # @param table [Symbol] the table name.
+  # @param params [Hash] the request parameters.
+  # @return [OpenStruct] validation results.
+  def validate_extended_columns(table, params)
+    validator = Crossbeams::Config::ExtendedColumnDefinitions::VALIDATIONS[table][AppConst::CLIENT_CODE]
+    return OpenStruct.new(messages: {}) unless validator
+
+    res = validator.call(select_extended_columns_params(params))
+    errs = { messages: res.messages.transform_keys { |k| "extcol_#{k}".to_sym } }
+    fields = res.to_h.transform_keys { |k| "extcol_#{k}".to_sym }
+    OpenStruct.new(errs.merge(fields))
+  end
+
+  # Add extended_columns to a changeset.
+  #
+  # @param changeset [Hash, DryStruct] the changeset.
+  # @param repo [BaseRepository] any repository.
+  # @param extended_cols [Hash] the extended_column values.
+  # @return [Hash] the augmented changeset.
+  def add_extended_columns_to_changeset(changeset, repo, extended_cols)
+    changeset.to_h.merge(extended_columns: repo.hash_for_jsonb_col(extended_cols))
+  end
+
+  # Get the extended_columns hash from an instance and change the keys from strings to symbols.
+  #
+  # @param instance [DryStruct/Hash] the data instance.
+  # @return [Hash] the extended_columns Hash or an empty Hash.
+  def extended_columns_for_row(instance)
+    return {} unless instance.to_h[:extended_columns]
+    instance.to_h[:extended_columns].symbolize_keys
+  end
+
   # Mark an entity as complete.
   #
   # @param table_name [string] the table.